# Js symbols continued (correctness and perf)