### Why?

1) There were correctness issues in cases where the js symbol file name contained `node:` prefixes.
2) Perf was poor.

### What?

Fix correctness and improve perf.

```
$ benchstat old.txt new.txt
name              old time/op  new time/op  delta
ParseJsSymbol-10   164ns ± 1%    43ns ± 2%  -73.64%  (p=0.008 n=5+5)
```

### How?

Remove allocations from parsing and use base 10 specific parsing function.

### Test Plan

Benchmark and unit tests.